Refactor C# integration to split it into multiple projects
