+++ integrations/dotnet/.gitignore
