+++ integrations/dotnet/js65.browser/js65.browser.csproj
@@ -1,0 +1,26 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFrameworks>net8.0-browser;net9.0-browser</TargetFrameworks>
+    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+    <OutputType>Library</OutputType>
+    <PackageId>js65.browser</PackageId>
+    <RootNamespace>js65</RootNamespace>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="../js65.interface/js65.interface.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <WasmExtraFilesToDeploy Include="$(ProjectDir)/../../../build/libassembler.js" TargetPath="js65/libassembler.js" />
+    <PackageReference Include="Microsoft.JSInterop" Version="8.0.11" />
+    <None Include="../js65/README.md" Pack="true" PackagePath="" />
+    <None Include="../js65/LICENSE" Pack="true" PackagePath="" />
+    <PackageReference Include="NuGetizer" Version="1.4.6">
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+      <PrivateAssets>all</PrivateAssets>
+    </PackageReference>
+  </ItemGroup>
+
+</Project>
