+++ integrations/dotnet/js65.clearscript/js65.clearscript.csproj
@@ -1,0 +1,26 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFrameworks>net8.0;net9.0</TargetFrameworks>
+    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+    <OutputType>Library</OutputType>
+    <PackageId>js65.clearscript</PackageId>
+    <RootNamespace>js65</RootNamespace>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="../js65.interface/js65.interface.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <EmbeddedResource Include="$(ProjectDir)/../../../build/libassembler.js" />
+    <PackageReference Include="Microsoft.ClearScript" Version="7.5.0" />
+    <None Include="../js65/README.md" Pack="true" PackagePath="" />
+    <None Include="../js65/LICENSE" Pack="true" PackagePath="" />
+    <PackageReference Include="NuGetizer" Version="1.4.6">
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+      <PrivateAssets>all</PrivateAssets>
+    </PackageReference>
+  </ItemGroup>
+
+</Project>
