+++ integrations/dotnet/js65.interface/AsmModule.cs
