+++ integrations/dotnet/js65.interface/Assembler.cs
