--- integrations/dotnet/js65/js65.csproj
@@ -1,34 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>net8.0;net8.0-browser</TargetFrameworks>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <OutputType>Library</OutputType>
-    <PackageId>js65.interop</PackageId>
-    <Version>1.0.0-alpha3</Version>
-    <Authors>Steve Hicks,James Rowe</Authors>
-    <PackageLicenseFile>LICENSE</PackageLicenseFile>
-    <PackageReadmeFile>README.md</PackageReadmeFile>
-  </PropertyGroup>
-
-  <ItemGroup Condition="'$(TargetFramework)' == 'net8.0'">
-    <Compile Remove="BrowserEngine.cs" />
-    <EmbeddedResource Include="$(ProjectDir)/../../../build/libassembler.js" />
-    <PackageReference Include="Microsoft.ClearScript" Version="7.5.0" />
-  </ItemGroup>
-  <ItemGroup Condition="'$(TargetFramework)' == 'net8.0-browser'">
-    <Compile Remove="ClearScriptEngine.cs" />
-    <PackageReference Include="Microsoft.JSInterop" Version="8.0.11" />
-    <WasmExtraFilesToDeploy Include="$(ProjectDir)/../../../build/libassembler.js" TargetPath="js65/libassembler.js" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="README.md" Pack="true" PackagePath="" />
-    <None Include="LICENSE" Pack="true" PackagePath="" />
-    <PackageReference Include="NuGetizer" Version="1.4.6">
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-      <PrivateAssets>all</PrivateAssets>
-    </PackageReference>
-  </ItemGroup>
-</Project>
